+++ Importer/ImportException.php
@@ -12,4 +12,13 @@ namespace KimaiPlugin\ImportBundle\Importer;
 
 class ImportException extends \Exception
 {
+    public function __construct(string $message, private readonly ?string $field = null)
+    {
+        parent::__construct($message);
+    }
+
+    public function getField(): ?string
+    {
+        return $this->field;
+    }
 }